Init expression type checking for defines, and const folding for casts.
The following define will now be reported as an error during compilation since the initialization expression has a different type than the define itself (float != int)

define float f1 = 100

Also, this changeset includes const folding for casts.  This optimization is always turned on for int, long, float, double casts.  If the -noOptimize flag is set, it will not be done for byte and short casts.  This is a good optimization to have in general, but it also makes the following define statement legal now:

define double timeout = (double)30sec