sox client: better type checking on add component
Issue 16868: allow better copying of objects from sax to sox

When adding a component using sox client, if the schema's didn't match
for the type and remote device, then you could not add the component even
if the kit parts were the same for that particular component.

This fix relaxes that restriction and will allow you to add a component to
a remote device so long as the type's kit part matches the remote device's
kit part.