fix java version checking bug in sedona.Env
Env.checkJavaVersion() assumed that the only non-digit, non-dot character
that would be present in the java.version system property would be '_'.
This is not the case, and it was causing bad warnings to be displayed
during compilation.

The fix uses the substring of the java.version property up to the first
non-valid character.