inet: prevent svm crash when ipv6 address used in udp socket
Issue 18436: opensource inet natives cause svm to crash when connecting to
ipv6 address.

This is a temporary fix to prevent the svm from crashing when an ipv6 address
is used for sending a datagram.  Our current inet implementation only supports
ipv4 sockets, but when using an ipv6 address it causeses a bad cast to an ipv6
struct with the result that unexpected memory locations get written.