sys: Fail when expected platform type does not exist
Issue 18615: App does not check for expected platform service correctly

In addition to this fix, the natives that return platformType() for win32
and unix have been updated to return sys::PlatformService.  Prior to this,
they returned an invalid service type. (Note - we cannot make them explicitly
return Win32PlatformService and UnixPlatformService because this will cause
the test suite to fail. Maybe something to look at later).