make Java sedona.Buf a random access buffer
Issue 18955: Buf does not properly handle random buffer access

The api for Buf was reworked to make it a truly random access
buffer. The methods were documented to better indicate behavior
of the read/write pointer after the method executed.

See also forum thread:
https://groups.google.com/forum/?fromgroups#!topic/sedonadev/C3RKDMgR7ms

Had to change one bit of code in the compiler (ImageGen) that used
some of the public fields in an unorthodox way to be compatible with
this change.  All other open source code appears to be ok.