dasp: drop challenge responses during authenticate phase of handshake
Issue 19144: Getting "Expected welcome, not 2" error when attempting
a sox connection.

Sometimes, when opening a sox connection, we get a DaspException
indicating "Expected welcome, not 2". In this case 2 means a challenge
response.  There are some scenarios in which we might get a challenge
message from a dasp server if multiple hello messages were sent
during that phase of the handshake.

So now we check the message type of the response before returning
from the authenticate phase of the handshake (on the client). We no
longer return challenge responses.