make this kit as a standalone git repo
the development on this kit will be move to git repo:
https://github.com/linsong/sedona-mqtt, this branch is just leaved here
for reference purpose from now on.

since this is a native kit, and to use it must recompile svm, ship it
together with sedona seems not appropriate, especially considering we
will have more native kits in future and user don't want every kit.

we can develop a kit management scipt that will help user to list,
search, install and uninstall other kit into his/her sedona folder.