support configurable url prefix
don't create StaticFileWeblet in static method '_sInit' like what
SpyWeblet does, let user to decide when to enable/disable it.

The current SpyWeblet is not possible to disable it unless uninstall the
web kit. I think we should modify it to be the same as StaticFileWeblet.
I will do it later to avoid this branch bloated.